DataEnhancement/backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
albucore==0.0.13
albumentations==1.4.10
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
attrs==25.3.0
beautifulsoup4==4.12.3
blinker==1.6.2
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
Crawl4AI==0.5.0.post8
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
distro==1.9.0
easyocr==1.7.2
et_xmlfile==2.0.0
fake-http-header==0.3.5
fake-useragent==2.1.0
fastapi==0.115.4
faust-cchardet==2.1.19
filelock==3.16.1
fire==0.7.0
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gspread==6.2.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.30.2
humanize==4.12.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
importlib_metadata==8.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.9.0
joblib==1.3.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
litellm==1.66.1
lmdb==1.5.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
narwhals==1.31.0
networkx==3.4.2
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.68.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.1
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pillow==10.4.0
playwright==1.51.0
propcache==0.3.1
protobuf==3.20.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.0
pyperclip==1.9.0
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.17
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.10.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.23.1
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.11.1
seaborn==0.13.2
selenium==4.30.0
selenium-stealth==1.0.6
shapely==2.0.6
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.2
streamlit==1.43.2
streamlit-cookies-controller==0.0.4
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
tf-playwright-stealth==1.1.2
threadpoolctl==3.2.0
tifffile==2024.9.20
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.66.6
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.27
ultralytics-thop==2.0.10
undetected-chromedriver==3.5.5
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.32.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.6
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
